+++ src/main/java/me/jellysquid/mods/sodium/client/render/chunk/compile/buffers/BakedChunkModelBuilder.java
@@ -1,15 +1,15 @@
 package me.jellysquid.mods.sodium.client.render.chunk.compile.buffers;
 
 import me.jellysquid.mods.sodium.client.model.quad.properties.ModelQuadFacing;
+import me.jellysquid.mods.sodium.client.render.chunk.data.BuiltSectionInfo;
 import me.jellysquid.mods.sodium.client.render.chunk.terrain.material.Material;
 import me.jellysquid.mods.sodium.client.render.chunk.vertex.builder.ChunkMeshBufferBuilder;
-import me.jellysquid.mods.sodium.client.render.chunk.data.BuiltSectionInfo;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.texture.Sprite;
 
 public class BakedChunkModelBuilder implements ChunkModelBuilder {
     private final ChunkMeshBufferBuilder[] vertexBuffers;
-    private final ChunkVertexConsumer vertexConsumerAdapter = new ChunkVertexConsumer(this);
+    private final ChunkVertexConsumer fallbackVertexConsumer = new ChunkVertexConsumer(this);
 
     private BuiltSectionInfo.Builder renderData;
 
@@ -28,9 +28,9 @@ public class BakedChunkModelBuilder implements ChunkModelBuilder {
     }
 
     @Override
-    public VertexConsumer asVertexConsumer(Material material) {
-        vertexConsumerAdapter.setMaterial(material);
-        return vertexConsumerAdapter;
+    public VertexConsumer asFallbackVertexConsumer(Material material) {
+        fallbackVertexConsumer.setMaterial(material);
+        return fallbackVertexConsumer;
     }
 
     public void destroy() {
