+++ src/main/java/me/jellysquid/mods/sodium/client/render/chunk/compile/pipeline/DefaultFluidRenderer.java
@@ -1,0 +1,459 @@
+package me.jellysquid.mods.sodium.client.render.chunk.compile.pipeline;
+
+import me.jellysquid.mods.sodium.client.model.color.ColorProvider;
+import me.jellysquid.mods.sodium.client.model.color.ColorProviderRegistry;
+import me.jellysquid.mods.sodium.client.model.color.DefaultColorProviders;
+import me.jellysquid.mods.sodium.client.model.light.LightMode;
+import me.jellysquid.mods.sodium.client.model.light.LightPipeline;
+import me.jellysquid.mods.sodium.client.model.light.LightPipelineProvider;
+import me.jellysquid.mods.sodium.client.model.light.data.QuadLightData;
+import me.jellysquid.mods.sodium.client.model.quad.ModelQuad;
+import me.jellysquid.mods.sodium.client.model.quad.ModelQuadView;
+import me.jellysquid.mods.sodium.client.model.quad.ModelQuadViewMutable;
+import me.jellysquid.mods.sodium.client.model.quad.properties.ModelQuadFacing;
+import me.jellysquid.mods.sodium.client.model.quad.properties.ModelQuadFlags;
+import me.jellysquid.mods.sodium.client.render.chunk.compile.buffers.ChunkModelBuilder;
+import me.jellysquid.mods.sodium.client.render.chunk.terrain.material.Material;
+import me.jellysquid.mods.sodium.client.render.chunk.vertex.format.ChunkVertexEncoder;
+import me.jellysquid.mods.sodium.client.util.DirectionUtil;
+import me.jellysquid.mods.sodium.client.world.WorldSlice;
+import net.caffeinemc.mods.sodium.api.util.ColorABGR;
+import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandler;
+import net.fabricmc.fabric.api.client.render.fluid.v1.FluidRenderHandlerRegistry;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.SideShapeType;
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.texture.Sprite;
+import net.minecraft.fluid.Fluid;
+import net.minecraft.fluid.FluidState;
+import net.minecraft.registry.tag.FluidTags;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.util.math.MathHelper;
+import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockRenderView;
+import org.apache.commons.lang3.mutable.MutableFloat;
+import org.apache.commons.lang3.mutable.MutableInt;
+
+public class DefaultFluidRenderer {
+    // TODO: allow this to be changed by vertex format
+    // TODO: move fluid rendering to a separate render pass and control glPolygonOffset and glDepthFunc to fix this properly
+    private static final float EPSILON = 0.001f;
+
+    private final ColorProviderRegistry colorProviderRegistry;
+    private final LightPipelineProvider lighters;
+
+    private final ModelQuadViewMutable quad = new ModelQuad();
+    private final QuadLightData quadLightData = new QuadLightData();
+    private final int[] quadColors = new int[4];
+    private final ChunkVertexEncoder.Vertex[] vertices = ChunkVertexEncoder.Vertex.uninitializedQuad();
+
+    private final BlockPos.Mutable scratchPos = new BlockPos.Mutable();
+    private final MutableFloat scratchHeight = new MutableFloat(0);
+    private final MutableInt scratchSamples = new MutableInt();
+
+    public DefaultFluidRenderer(ColorProviderRegistry colorProviderRegistry, LightPipelineProvider lighters) {
+        this.colorProviderRegistry = colorProviderRegistry;
+        this.lighters = lighters;
+        this.quad.setLightFace(Direction.UP);
+    }
+
+    private boolean isFluidOccluded(BlockRenderView world, int x, int y, int z, Direction dir, Fluid fluid) {
+        BlockPos pos = this.scratchPos.set(x, y, z);
+        BlockState blockState = world.getBlockState(pos);
+        BlockPos adjPos = this.scratchPos.set(x + dir.getOffsetX(), y + dir.getOffsetY(), z + dir.getOffsetZ());
+
+        if (blockState.isOpaque()) {
+            return world.getFluidState(adjPos).getFluid().matchesType(fluid) || blockState.isSideSolid(world, pos, dir, SideShapeType.FULL);
+        }
+        return world.getFluidState(adjPos).getFluid().matchesType(fluid);
+    }
+
+    private boolean isSideExposed(BlockRenderView world, int x, int y, int z, Direction dir, float height) {
+        BlockPos pos = this.scratchPos.set(x + dir.getOffsetX(), y + dir.getOffsetY(), z + dir.getOffsetZ());
+        BlockState blockState = world.getBlockState(pos);
+
+        if (blockState.isOpaque()) {
+            VoxelShape shape = blockState.getCullingShape(world, pos);
+
+            // Hoist these checks to avoid allocating the shape below
+            if (shape == VoxelShapes.fullCube()) {
+                // The top face always be inset, so if the shape above is a full cube it can't possibly occlude
+                return dir == Direction.UP;
+            } else if (shape.isEmpty()) {
+                return true;
+            }
+
+            VoxelShape threshold = VoxelShapes.cuboid(0.0D, 0.0D, 0.0D, 1.0D, height, 1.0D);
+
+            return !VoxelShapes.isSideCovered(threshold, shape, dir);
+        }
+
+        return true;
+    }
+
+    public void render(WorldSlice world, FluidState fluidState, BlockPos blockPos, BlockPos offset, ChunkModelBuilder meshBuilder, Material material, FluidRenderHandler handler) {
+        int posX = blockPos.getX();
+        int posY = blockPos.getY();
+        int posZ = blockPos.getZ();
+
+        Fluid fluid = fluidState.getFluid();
+
+        boolean sfUp = this.isFluidOccluded(world, posX, posY, posZ, Direction.UP, fluid);
+        boolean sfDown = this.isFluidOccluded(world, posX, posY, posZ, Direction.DOWN, fluid) ||
+                !this.isSideExposed(world, posX, posY, posZ, Direction.DOWN, 0.8888889F);
+        boolean sfNorth = this.isFluidOccluded(world, posX, posY, posZ, Direction.NORTH, fluid);
+        boolean sfSouth = this.isFluidOccluded(world, posX, posY, posZ, Direction.SOUTH, fluid);
+        boolean sfWest = this.isFluidOccluded(world, posX, posY, posZ, Direction.WEST, fluid);
+        boolean sfEast = this.isFluidOccluded(world, posX, posY, posZ, Direction.EAST, fluid);
+
+        if (sfUp && sfDown && sfEast && sfWest && sfNorth && sfSouth) {
+            return;
+        }
+
+        Sprite[] sprites = handler.getFluidSprites(world, blockPos, fluidState);
+        final ColorProvider<FluidState> colorProvider = this.getColorProvider(fluid, handler);
+
+        float fluidHeight = this.fluidHeight(world, fluid, blockPos, Direction.UP);
+        float northWestHeight, southWestHeight, southEastHeight, northEastHeight;
+        if (fluidHeight >= 1.0f) {
+            northWestHeight = 1.0f;
+            southWestHeight = 1.0f;
+            southEastHeight = 1.0f;
+            northEastHeight = 1.0f;
+        } else {
+            var scratchPos = new BlockPos.Mutable();
+            float heightNorth = this.fluidHeight(world, fluid, scratchPos.set(blockPos, Direction.NORTH), Direction.NORTH);
+            float heightSouth = this.fluidHeight(world, fluid, scratchPos.set(blockPos, Direction.SOUTH), Direction.SOUTH);
+            float heightEast = this.fluidHeight(world, fluid, scratchPos.set(blockPos, Direction.EAST), Direction.EAST);
+            float heightWest = this.fluidHeight(world, fluid, scratchPos.set(blockPos, Direction.WEST), Direction.WEST);
+            northWestHeight = this.fluidCornerHeight(world, fluid, fluidHeight, heightNorth, heightWest, scratchPos.set(blockPos)
+                    .move(Direction.NORTH)
+                    .move(Direction.WEST));
+            southWestHeight = this.fluidCornerHeight(world, fluid, fluidHeight, heightSouth, heightWest, scratchPos.set(blockPos)
+                    .move(Direction.SOUTH)
+                    .move(Direction.WEST));
+            southEastHeight = this.fluidCornerHeight(world, fluid, fluidHeight, heightSouth, heightEast, scratchPos.set(blockPos)
+                    .move(Direction.SOUTH)
+                    .move(Direction.EAST));
+            northEastHeight = this.fluidCornerHeight(world, fluid, fluidHeight, heightNorth, heightEast, scratchPos.set(blockPos)
+                    .move(Direction.NORTH)
+                    .move(Direction.EAST));
+        }
+        float yOffset = sfDown ? 0.0F : EPSILON;
+
+        final ModelQuadViewMutable quad = this.quad;
+
+        LightMode lightMode = fluidState.isIn(FluidTags.WATER) && MinecraftClient.isAmbientOcclusionEnabled() ? LightMode.SMOOTH : LightMode.FLAT;
+        LightPipeline lighter = this.lighters.getLighter(lightMode);
+
+        quad.setFlags(0);
+
+        if (!sfUp && this.isSideExposed(world, posX, posY, posZ, Direction.UP, Math.min(Math.min(northWestHeight, southWestHeight), Math.min(southEastHeight, northEastHeight)))) {
+            northWestHeight -= EPSILON;
+            southWestHeight -= EPSILON;
+            southEastHeight -= EPSILON;
+            northEastHeight -= EPSILON;
+
+            Vec3d velocity = fluidState.getVelocity(world, blockPos);
+
+            Sprite sprite;
+            ModelQuadFacing facing;
+            float u1, u2, u3, u4;
+            float v1, v2, v3, v4;
+
+            if (velocity.x == 0.0D && velocity.z == 0.0D) {
+                sprite = sprites[0];
+                facing = ModelQuadFacing.POS_Y;
+                u1 = sprite.getFrameU(0.0f);
+                v1 = sprite.getFrameV(0.0f);
+                u2 = u1;
+                v2 = sprite.getFrameV(1.0f);
+                u3 = sprite.getFrameU(1.0f);
+                v3 = v2;
+                u4 = u3;
+                v4 = v1;
+            } else {
+                sprite = sprites[1];
+                facing = ModelQuadFacing.UNASSIGNED;
+                float dir = (float) MathHelper.atan2(velocity.z, velocity.x) - (1.5707964f);
+                float sin = MathHelper.sin(dir) * 0.25F;
+                float cos = MathHelper.cos(dir) * 0.25F;
+                u1 = sprite.getFrameU(0.5F + (-cos - sin));
+                v1 = sprite.getFrameV(0.5F + -cos + sin);
+                u2 = sprite.getFrameU(0.5F + -cos + sin);
+                v2 = sprite.getFrameV(0.5F + cos + sin);
+                u3 = sprite.getFrameU(0.5F + cos + sin);
+                v3 = sprite.getFrameV(0.5F + (cos - sin));
+                u4 = sprite.getFrameU(0.5F + (cos - sin));
+                v4 = sprite.getFrameV(0.5F + (-cos - sin));
+            }
+
+            float uAvg = (u1 + u2 + u3 + u4) / 4.0F;
+            float vAvg = (v1 + v2 + v3 + v4) / 4.0F;
+            float s3 = sprites[0].getAnimationFrameDelta();
+
+            u1 = MathHelper.lerp(s3, u1, uAvg);
+            u2 = MathHelper.lerp(s3, u2, uAvg);
+            u3 = MathHelper.lerp(s3, u3, uAvg);
+            u4 = MathHelper.lerp(s3, u4, uAvg);
+            v1 = MathHelper.lerp(s3, v1, vAvg);
+            v2 = MathHelper.lerp(s3, v2, vAvg);
+            v3 = MathHelper.lerp(s3, v3, vAvg);
+            v4 = MathHelper.lerp(s3, v4, vAvg);
+
+            quad.setSprite(sprite);
+
+            setVertex(quad, 0, 0.0f, northWestHeight, 0.0f, u1, v1);
+            setVertex(quad, 1, 0.0f, southWestHeight, 1.0F, u2, v2);
+            setVertex(quad, 2, 1.0F, southEastHeight, 1.0F, u3, v3);
+            setVertex(quad, 3, 1.0F, northEastHeight, 0.0f, u4, v4);
+
+            this.updateQuad(quad, world, blockPos, lighter, Direction.UP, 1.0F, colorProvider, fluidState);
+            this.writeQuad(meshBuilder, material, offset, quad, facing, false);
+
+            if (fluidState.canFlowTo(world, this.scratchPos.set(posX, posY + 1, posZ))) {
+                this.writeQuad(meshBuilder, material, offset, quad,
+                        ModelQuadFacing.NEG_Y, true);
+            }
+        }
+
+        if (!sfDown) {
+            Sprite sprite = sprites[0];
+
+            float minU = sprite.getMinU();
+            float maxU = sprite.getMaxU();
+            float minV = sprite.getMinV();
+            float maxV = sprite.getMaxV();
+            quad.setSprite(sprite);
+
+            setVertex(quad, 0, 0.0f, yOffset, 1.0F, minU, maxV);
+            setVertex(quad, 1, 0.0f, yOffset, 0.0f, minU, minV);
+            setVertex(quad, 2, 1.0F, yOffset, 0.0f, maxU, minV);
+            setVertex(quad, 3, 1.0F, yOffset, 1.0F, maxU, maxV);
+
+            this.updateQuad(quad, world, blockPos, lighter, Direction.DOWN, 1.0F, colorProvider, fluidState);
+            this.writeQuad(meshBuilder, material, offset, quad, ModelQuadFacing.NEG_Y, false);
+        }
+
+        quad.setFlags(ModelQuadFlags.IS_PARALLEL | ModelQuadFlags.IS_ALIGNED);
+
+        for (Direction dir : DirectionUtil.HORIZONTAL_DIRECTIONS) {
+            float c1;
+            float c2;
+            float x1;
+            float z1;
+            float x2;
+            float z2;
+
+            switch (dir) {
+                case NORTH -> {
+                    if (sfNorth) {
+                        continue;
+                    }
+                    c1 = northWestHeight;
+                    c2 = northEastHeight;
+                    x1 = 0.0f;
+                    x2 = 1.0F;
+                    z1 = EPSILON;
+                    z2 = z1;
+                }
+                case SOUTH -> {
+                    if (sfSouth) {
+                        continue;
+                    }
+                    c1 = southEastHeight;
+                    c2 = southWestHeight;
+                    x1 = 1.0F;
+                    x2 = 0.0f;
+                    z1 = 1.0f - EPSILON;
+                    z2 = z1;
+                }
+                case WEST -> {
+                    if (sfWest) {
+                        continue;
+                    }
+                    c1 = southWestHeight;
+                    c2 = northWestHeight;
+                    x1 = EPSILON;
+                    x2 = x1;
+                    z1 = 1.0F;
+                    z2 = 0.0f;
+                }
+                case EAST -> {
+                    if (sfEast) {
+                        continue;
+                    }
+                    c1 = northEastHeight;
+                    c2 = southEastHeight;
+                    x1 = 1.0f - EPSILON;
+                    x2 = x1;
+                    z1 = 0.0f;
+                    z2 = 1.0F;
+                }
+                default -> {
+                    continue;
+                }
+            }
+
+            if (this.isSideExposed(world, posX, posY, posZ, dir, Math.max(c1, c2))) {
+                int adjX = posX + dir.getOffsetX();
+                int adjY = posY + dir.getOffsetY();
+                int adjZ = posZ + dir.getOffsetZ();
+
+                Sprite sprite = sprites[1];
+
+                boolean isOverlay = false;
+
+                if (sprites.length > 2) {
+                    BlockPos adjPos = this.scratchPos.set(adjX, adjY, adjZ);
+                    BlockState adjBlock = world.getBlockState(adjPos);
+
+                    if (FluidRenderHandlerRegistry.INSTANCE.isBlockTransparent(adjBlock.getBlock())) {
+                        sprite = sprites[2];
+                        isOverlay = true;
+                    }
+                }
+
+                float u1 = sprite.getFrameU(0.0F);
+                float u2 = sprite.getFrameU(0.5F);
+                float v1 = sprite.getFrameV((1.0F - c1) * 0.5F);
+                float v2 = sprite.getFrameV((1.0F - c2) * 0.5F);
+                float v3 = sprite.getFrameV(0.5F);
+
+                quad.setSprite(sprite);
+
+                setVertex(quad, 0, x2, c2, z2, u2, v2);
+                setVertex(quad, 1, x2, yOffset, z2, u2, v3);
+                setVertex(quad, 2, x1, yOffset, z1, u1, v3);
+                setVertex(quad, 3, x1, c1, z1, u1, v1);
+
+                float br = dir.getAxis() == Direction.Axis.Z ? 0.8F : 0.6F;
+
+                ModelQuadFacing facing = ModelQuadFacing.fromDirection(dir);
+
+                this.updateQuad(quad, world, blockPos, lighter, dir, br, colorProvider, fluidState);
+                this.writeQuad(meshBuilder, material, offset, quad, facing, false);
+
+                if (!isOverlay) {
+                    this.writeQuad(meshBuilder, material, offset, quad, facing.getOpposite(), true);
+                }
+            }
+        }
+    }
+
+    private ColorProvider<FluidState> getColorProvider(Fluid fluid, FluidRenderHandler handler) {
+        var override = this.colorProviderRegistry.getColorProvider(fluid);
+
+        if (override != null) {
+            return override;
+        }
+        
+        return DefaultColorProviders.adapt(handler);
+    }
+
+    private void updateQuad(ModelQuadView quad, WorldSlice world, BlockPos pos, LightPipeline lighter, Direction dir, float brightness,
+                            ColorProvider<FluidState> colorProvider, FluidState fluidState) {
+        QuadLightData light = this.quadLightData;
+        lighter.calculate(quad, pos, light, null, dir, false);
+
+        colorProvider.getColors(world, pos, fluidState, quad, this.quadColors);
+
+        // multiply the per-vertex color against the combined brightness
+        // the combined brightness is the per-vertex brightness multiplied by the block's brightness
+        for (int i = 0; i < 4; i++) {
+            this.quadColors[i] = ColorABGR.withAlpha(this.quadColors[i], light.br[i] * brightness);
+        }
+    }
+
+    private void writeQuad(ChunkModelBuilder builder, Material material, BlockPos offset, ModelQuadView quad,
+                           ModelQuadFacing facing, boolean flip) {
+        var vertices = this.vertices;
+
+        for (int i = 0; i < 4; i++) {
+            var out = vertices[flip ? 3 - i : i];
+            out.x = offset.getX() + quad.getX(i);
+            out.y = offset.getY() + quad.getY(i);
+            out.z = offset.getZ() + quad.getZ(i);
+
+            out.color = this.quadColors[i];
+            out.u = quad.getTexU(i);
+            out.v = quad.getTexV(i);
+            out.light = this.quadLightData.lm[i];
+        }
+
+        Sprite sprite = quad.getSprite();
+
+        if (sprite != null) {
+            builder.addSprite(sprite);
+        }
+
+        var vertexBuffer = builder.getVertexBuffer(facing);
+        vertexBuffer.push(vertices, material);
+    }
+
+    private static void setVertex(ModelQuadViewMutable quad, int i, float x, float y, float z, float u, float v) {
+        quad.setX(i, x);
+        quad.setY(i, y);
+        quad.setZ(i, z);
+        quad.setTexU(i, u);
+        quad.setTexV(i, v);
+    }
+
+    private float fluidCornerHeight(BlockRenderView world, Fluid fluid, float fluidHeight, float fluidHeightX, float fluidHeightY, BlockPos blockPos) {
+        if (fluidHeightY >= 1.0f || fluidHeightX >= 1.0f) {
+            return 1.0f;
+        }
+
+        if (fluidHeightY > 0.0f || fluidHeightX > 0.0f) {
+            float height = this.fluidHeight(world, fluid, blockPos, Direction.UP);
+
+            if (height >= 1.0f) {
+                return 1.0f;
+            }
+
+            this.modifyHeight(this.scratchHeight, this.scratchSamples, height);
+        }
+
+        this.modifyHeight(this.scratchHeight, this.scratchSamples, fluidHeight);
+        this.modifyHeight(this.scratchHeight, this.scratchSamples, fluidHeightY);
+        this.modifyHeight(this.scratchHeight, this.scratchSamples, fluidHeightX);
+
+        float result = this.scratchHeight.floatValue() / this.scratchSamples.intValue();
+        this.scratchHeight.setValue(0);
+        this.scratchSamples.setValue(0);
+
+        return result;
+    }
+
+    private void modifyHeight(MutableFloat totalHeight, MutableInt samples, float target) {
+        if (target >= 0.8f) {
+            totalHeight.add(target * 10.0f);
+            samples.add(10);
+        } else if (target >= 0.0f) {
+            totalHeight.add(target);
+            samples.increment();
+        }
+    }
+
+    private float fluidHeight(BlockRenderView world, Fluid fluid, BlockPos blockPos, Direction direction) {
+        BlockState blockState = world.getBlockState(blockPos);
+        FluidState fluidState = blockState.getFluidState();
+
+        if (fluid.matchesType(fluidState.getFluid())) {
+            FluidState fluidStateUp = world.getFluidState(blockPos.up());
+
+            if (fluid.matchesType(fluidStateUp.getFluid())) {
+                return 1.0f;
+            } else {
+                return fluidState.getHeight();
+            }
+        }
+        if (!blockState.isSolid()) {
+            return 0.0f;
+        }
+        return -1.0f;
+    }
+}
