+++ src/main/resources/fabric.mod.json
@@ -73,7 +73,7 @@
     "fabric-block-view-api-v2": "*",
     "fabric-renderer-api-v1": "*",
     "fabric-rendering-data-attachment-v1": "*",
-    "fabric-rendering-fluids-v1": "*",
+    "fabric-rendering-fluids-v1": ">=2.0.0",
     "fabric-resource-loader-v0": "*"
   },
   "breaks": {
